refactor: derive unused internal addresses via hook
